fix cupertino destroy on every nav

+++ src/lib/stores/cupertino-pane/cupertino-pane.store.ts
@@ -49,6 +49,8 @@ export default (() => {
   }
 
   function closeSheet() {
+    if (!get(store).component) return;
+
     const p = get(pane);
 
     p.destroy({ animate: true });
